Refactor into Java from Kotlin I wanted to try Kotlin out, and it turns out I absolutely hate it and never want to use it again.
Much of this code has been ported over from the previous project - https://github.com/MarkStrendin/MSPwdGen_Android, to ensure that it works the same, and that passwords created with the old program are the same as those created with this one.

At this point the app appears to work in the emulator, so its just in need of some tidying up and polishing.